# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Shaghlny - Job Hiring Platform API",
        "description": "API documentation for Shaghlny - Job Hiring Platform",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/company/position/create": {
            "get": {
                "tags": [
                    "Company"
                ],
                "summary": "Create a new job position",
                "description": "Returns a message instructing the user on how to submit a job position.",
                "operationId": "0c6cc1df8bdcff492ff7c1c805979283",
                "responses": {
                    "200": {
                        "description": "Position creation instructions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Use this endpoint to submit position details."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/company/position": {
            "post": {
                "tags": [
                    "Company"
                ],
                "summary": "Store a new job position with associated questions",
                "description": "Store a new job position with a title, description, location, and optional questions.",
                "operationId": "be14d0154c6e1e14412e15098bc2c5d9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Software Engineer"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Develop and maintain software."
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "New York, NY"
                                    },
                                    "questions": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "What is your experience with PHP?",
                                            "How do you handle deadlines?"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Job position and questions created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Position and questions added successfully!"
                                        },
                                        "job": {
                                            "$ref": "#/components/schemas/Job"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Job": {
                "required": [
                    "title",
                    "description",
                    "job_type"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "title": {
                        "type": "string",
                        "example": "Software Developer"
                    },
                    "description": {
                        "type": "string",
                        "example": "Job description here..."
                    },
                    "location": {
                        "type": "string",
                        "example": "Remote"
                    },
                    "salary": {
                        "type": "number",
                        "format": "float",
                        "example": 50000
                    },
                    "job_type": {
                        "type": "string",
                        "enum": [
                            "Full-time",
                            "Part-time",
                            "Contract"
                        ],
                        "example": "Full-time"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-11-25T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-11-25T12:00:00Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}